web/requirements.txt
anyio==3.6.1
async-generator==1.10
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==21.4.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
databases==0.6.0
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
identify==2.5.1
idna==3.3
iniconfig==1.1.1
minio==7.1.9
mypy-extensions==0.4.3
nodeenv==1.7.0
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
trio==0.21.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
